working on leave request add completed wokring on edit

+++ work-smart-backend/aircore/src/main/java/com/aircore/entity/User.java
@@ -54,7 +54,7 @@ public class User {
     
     private Date createdDate;
     
-    private Long totalLeave;
+    private Double totalLeave;
     
     @ManyToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
     @JoinTable(
